bugs fix

+++ 0x0B-malloc_free/3-alloc_grid.c
@@ -21,7 +21,7 @@ if (s == NULL)
 
 
 
-for (i = 0; i < (unsigned int)height; i++)
+for (i = 0; i < height; i++)
 {
 	s[i] = (int *)malloc(width * sizeof(int));
 	if (s[i] == NULL)
